[feat] 나의 시간표 조회 api 연결

+++ app/src/main/java/com/sopt/gongbaek/data/remote/dto/request/RegisterTimeTableRequestDto.kt
@@ -8,7 +8,7 @@ data class RegisterTimeTableRequestDto(
     @SerialName("weekDay")
     val weekDay: String,
     @SerialName("startTime")
-    val startTime: Float,
+    val startTime: Double,
     @SerialName("endTime")
-    val endTime: Float
+    val endTime: Double
 )
